@@ -22,7 +22,7 @@
     "file-loader": "0.8.4",
     "json-loader": "0.5.2",
     "node-libs-browser": "0.5.2",
-    "node-sass": "3.2.0",
+    "node-sass": "^4.5.1",
     "raw-loader": "0.5.1",
     "style-loader": "0.12.3",
     "url-loader": "0.5.6",
@@ -31,11 +31,12 @@
   },
   "dependencies": {
     "angular": "1.4.1",
-    "eventemitter2": "^0.4.14",
-    "flux": "^2.1.1",
-    "keymirror": "^0.1.1",
-    "lodash": "^3.10.1",
-    "object-assign": "^4.0.1",
-    "react": "^0.13.3"
+    "eventemitter2": "0.4.14",
+    "flux": "2.1.1",
+    "keymirror": "0.1.1",
+    "lodash": "3.10.1",
+    "object-assign": "4.0.1",
+    "react": "15.4.2",
+    "react-dom": "15.4.2"
   }
 }
